Implement crypto UI!
What:

1) Removed 'Assets' header, created headers for 'Stocks' and
'Cryptcurrencies'
2) Added v-if section for crypto, mimicking existing stocks setup
3) Fixed bug in AJAX where wrong jsonb property name was being used
4) Fixed bug in navigation_drawer where if you scrolled before clicking
dropdown menu, it wouldn't show the contents of it (the contents was
also scrolling)

Why:

1) Better UX
2) Needed
3) Needed. NOTE: The response from crypto api has duplicate values when
accessing US market. For example you'll get a key of "4a. close (USD)"
and another key of "4b. close (USD). Best to use the first one, incase
we want to implement other markets later as that is the one that will
change if select a different market. The second USD one is for
comparison
4) Needed. Glad I found this